# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artist Booking API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quotes": {
      "post": {
        "tags": [
          "quotes-v2",
          "QuotesV2"
        ],
        "summary": "Create Quote",
        "operationId": "create_quote_api_v1_quotes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__quote_v2__QuoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{quote_id}": {
      "get": {
        "tags": [
          "quotes",
          "Quotes"
        ],
        "summary": "Read Quote",
        "description": "Retrieve a specific quote by ID.\nAccessible by the client of the booking request or the artist who made the quote.",
        "operationId": "read_quote_api_v1_quotes__quote_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{quote_id}/accept": {
      "post": {
        "tags": [
          "quotes-v2",
          "QuotesV2"
        ],
        "summary": "Accept Quote",
        "operationId": "accept_quote_api_v1_quotes__quote_id__accept_post",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSimpleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{quote_id}/decline": {
      "post": {
        "tags": [
          "quotes-v2",
          "QuotesV2"
        ],
        "summary": "Decline Quote",
        "operationId": "decline_quote_api_v1_quotes__quote_id__decline_post",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{quote_id}/pdf": {
      "get": {
        "tags": [
          "quotes-v2",
          "QuotesV2"
        ],
        "summary": "Get Quote Pdf",
        "operationId": "get_quote_pdf_api_v1_quotes__quote_id__pdf_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingSimpleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "quote_id": {
            "type": "integer",
            "title": "Quote Id"
          },
          "artist_id": {
            "type": "integer",
            "title": "Artist Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "quote_id",
          "artist_id",
          "client_id",
          "confirmed",
          "payment_status",
          
          "created_at",
          "updated_at"
        ],
        "title": "BookingSimpleRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuoteRead": {
        "properties": {
          "booking_request_id": {
            "type": "integer",
            "title": "Booking Request Id"
          },
          "artist_id": {
            "type": "integer",
            "title": "Artist Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceItem-Output"
            },
            "type": "array",
            "title": "Services"
          },
          "sound_fee": {
            "type": "string",
            "title": "Sound Fee",
            "default": 0
          },
          "travel_fee": {
            "type": "string",
            "title": "Travel Fee",
            "default": 0
          },
          "accommodation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation"
          },
          "discount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "booking_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "subtotal": {
            "type": "string",
            "title": "Subtotal"
          },
          "total": {
            "type": "string",
            "title": "Total"
          },
          "status": {
            "$ref": "#/components/schemas/QuoteStatusV2"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "booking_request_id",
          "artist_id",
          "client_id",
          "services",
          "id",
          "subtotal",
          "total",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "QuoteRead"
      },
      "QuoteResponse": {
        "properties": {
          "quote_details": {
            "type": "string",
            "title": "Quote Details"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "ZAR"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "booking_request_id": {
            "type": "integer",
            "title": "Booking Request Id"
          },
          "artist_id": {
            "type": "integer",
            "title": "Artist Id"
          },
          "status": {
            "$ref": "#/components/schemas/QuoteStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "artist": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "artist_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtistProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "quote_details",
          "price",
          "id",
          "booking_request_id",
          "artist_id",
          "status",
          "created_at"
        ],
        "title": "QuoteResponse"
      },
      "QuoteStatus": {
        "type": "string",
        "enum": [
          "pending_client_action",
          "accepted_by_client",
          "rejected_by_client",
          "confirmed_by_artist",
          "withdrawn_by_artist",
          "expired"
        ],
        "title": "QuoteStatus"
      },
      "QuoteStatusV2": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "rejected",
          "expired"
        ],
        "title": "QuoteStatusV2"
      },
...
      "ServiceItem-Input": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "description",
          "price"
        ],
        "title": "ServiceItem"
      },
      "ServiceItem-Output": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "type": "string",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "description",
          "price"
        ],
        "title": "ServiceItem"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "mfa_enabled": {
            "type": "boolean",
            "title": "Mfa Enabled"
          },
          "profile_picture_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Picture Url"
          }
        },
        "type": "object",
        "required": [
          "email",
          "first_name",
          "last_name",
          "phone_number",
          "user_type",
          "id",
          "is_active",
          "is_verified",
          "mfa_enabled"
        ],
        "title": "UserResponse"
      },
      "UserType": {
        "type": "string",
        "enum": [
          "service_provider",
          "client"
        ],
        "title": "UserType",
        "description": "Roles supported by the API."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__quote_v2__QuoteCreate": {
        "properties": {
          "booking_request_id": {
            "type": "integer",
            "title": "Booking Request Id"
          },
          "artist_id": {
            "type": "integer",
            "title": "Artist Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceItem-Input"
            },
            "type": "array",
            "title": "Services"
          },
          "sound_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Sound Fee",
            "default": 0
          },
          "travel_fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Travel Fee",
            "default": 0
          },
          "accommodation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation"
          },
          "discount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "booking_request_id",
          "artist_id",
          "client_id",
          "services"
        ],
        "title": "QuoteCreate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}